pyproject.toml
[tool.poetry]
name = "data-analysis-bot"
version = "0.1.0"
description = ""
authors = ["Kai Erik Niermann <kai.niermann@gmail.com>"]
readme = "README.md"
packages = [{include = "./src"}]

[tool.poetry.dependencies]
python = ">=3.8.6,<4.0"
aiohttp = "3.8.4"
aiosignal = "1.3.1"
async-generator = "1.10"
async-timeout = "4.0.2"
attrs = "22.2.0"
autopep8 = "2.0.2"
black = "23.1.0"
certifi = "2022.12.7"
charset-normalizer = "3.1.0"
chromedriver-py = "112.0.5615.28"
click = "8.1.3"
contourpy = "1.0.7"
cycler = "0.11.0"
discord = "2.2.2"
discord-py-slash-command = "4.2.1"
exceptiongroup = "1.1.1"
ffmpeg-python = "0.2.0"
fonttools = "4.39.2"
frozenlist = "1.3.3"
future = "0.18.3"
h11 = "0.14.0"
idna = "3.4"
imageio-ffmpeg = "0.4.8"
importlib-resources = "5.12.0"
kiwisolver = "1.4.4"
markovify = "0.9.4"
matplotlib = "3.7.1"
multidict = "6.0.4"
mypy-extensions = "1.0.0"
numpy = "1.24.2"
outcome = "1.2.0"
packaging = "23.0"
pathspec = "0.11.1"
pillow = "9.4.0"
platformdirs = "3.1.1"
py-cord = "2.4.0"
pycodestyle = "2.10.0"
pyparsing = "3.0.9"
pysocks = "1.7.1"
python-dateutil = "2.8.2"
python-dotenv = "1.0.0"
pyyaml = "6.0"
requests = "2.28.2"
selenium = "4.8.2"
six = "1.16.0"
sniffio = "1.3.0"
sortedcontainers = "2.4.0"
tomli = "2.0.1"
tqdm = "4.65.0"
trio = "0.22.0"
trio-websocket = "0.10.1"
typing-extensions = "4.5.0"
unidecode = "1.3.6"
urllib3 = "1.26.15"
webdriver-manager = "3.8.5"
wsproto = "1.2.0"
yarl = "1.8.2"
youtube-dl = "2021.12.17"
zipp = "3.15.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
